add template engine
