+++ src/Argon.Api/Program.cs
@@ -6,12 +6,12 @@ using Argon.Api.Entities;
 using Argon.Api.Extensions;
 using Argon.Api.Features;
 using Argon.Api.Features.Captcha;
-using Argon.Api.Features.EmailForms;
 using Argon.Api.Features.Env;
 using Argon.Api.Features.Jwt;
 using Argon.Api.Features.MediaStorage;
 using Argon.Api.Features.Otp;
 using Argon.Api.Features.Pex;
+using Argon.Api.Features.Template;
 using Argon.Api.Grains.Interfaces;
 using Argon.Api.Migrations;
 using Argon.Api.Services;
@@ -49,7 +49,7 @@ builder.Services.AddTransient<UserManagerService>();
 builder.Services.AddSingleton<IFusionContext, FusionContext>();
 builder.AddOtpCodes();
 builder.AddOrleans();
-builder.AddEMailForms();
+builder.AddTemplateEngine();
 builder.AddKubeResources();
 builder.AddCaptchaFeature();
 builder.Services.AddDataProtection();
